pyproject.toml
[project]
name = "dran"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-queryfields>=1.1.0",
    "drf-nested-routers>=0.94.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "django-stubs>=5.1.3",
    "djangorestframework-stubs>=3.15.3",
    "mypy>=1.15.0",
]

[tool.ruff]
line-length = 119

[tool.mypy]
# Coverage
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
warn_incomplete_stub = true

# Any and Generics
disallow_any_decorated = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
warn_return_any = true

# Something unused
warn_unused_configs = true
warn_unused_ignores = true

# Possible errors
disallow_any_unimported = true
warn_redundant_casts = true
strict_equality = true
strict_concatenate = true
warn_unreachable = true
no_implicit_reexport = true

# Extra
strict = true

# django and drf stubs setup
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "blog.blog"

# Voodoo
[[tool.mypy.overrides]]
module = "blog.manage"
disallow_untyped_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "blog.api.serializers"
disallow_any_unimported = false

[[tool.mypy.overrides]]
module = "blog.api.permissions"
disallow_any_explicit = false

[[tool.mypy.overrides]]
module = "blog.api.views"
disable_error_code = "attr-defined"
